install promised-neovim-client because of global.require().
neovim-component invokes global.require('promised-neovim-client')
internally but module.paths are set by Electron render process.  It
causes 'no module' error on the expression because the
promised-neovim-client module path is not included module.path.
This commit avoids it by installing promised-neovim-client to root
node_modules.

This should be workaround for npm2.  npm3 should not occur the error
because of flat module directories layout.